[nft-transfer] fix unused warnings when compiling test with runtime-benchmarks
